truetandem/e-QIP-prototype#480 Adjust gaps to start at end of month rather than the first day
